Binance fix

+++ project/OsEngine/Market/Servers/Binance/Spot/BinanceSpotEntity/TradeResponse.cs
@@ -5,11 +5,11 @@
         public string e { get; set; }
         public long E { get; set; }
         public string s { get; set; }
-        public int t { get; set; }
+        public long t { get; set; }
         public string p { get; set; }
         public string q { get; set; }
-        public int b { get; set; }
-        public int a { get; set; }
+        public long b { get; set; }
+        public long a { get; set; }
         public long T { get; set; }
         public bool m { get; set; }
         public bool M { get; set; }
